Replace impact function in EMEnderPearl with Consumer
