+++ src/main/java/ru/mousecray/endmagic/items/BlueEnderPearl.java
@@ -1,29 +1,17 @@
 package ru.mousecray.endmagic.items;
 
 import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.projectile.EntityThrowable;
-import net.minecraft.init.SoundEvents;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.stats.StatList;
-import net.minecraft.util.ActionResult;
 import net.minecraft.util.DamageSource;
-import net.minecraft.util.EnumActionResult;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.world.World;
-import ru.mousecray.endmagic.entity.EntityEMEnderPearl;
-import ru.mousecray.endmagic.util.EntityImpact;
 
-public class BlueEnderPearl extends Item implements EntityImpact,ItemOneWhiteEMTextured {
+public class BlueEnderPearl extends EMEnderPearl {
 
 	public BlueEnderPearl() {
 		setMaxStackSize(16);
 	}
 	
 	@Override
-	public void onImpact(EntityLivingBase result, EntityThrowable entity, EntityLivingBase thrower) {
+	public void onImpact(EntityLivingBase result, EntityLivingBase thrower, EntityThrowable trowable) {
 		double posX = result.posX;
 		double posY = result.posY;
 		double posZ = result.posZ;
@@ -36,25 +24,6 @@ public class BlueEnderPearl extends Item implements EntityImpact,ItemOneWhiteEMT
 		thrower.fallDistance = 0.0F;
 		thrower.attackEntityFrom(DamageSource.FALL, 5F);
 	}
-	
-	@Override
-    public ActionResult<ItemStack> onItemRightClick(World world, EntityPlayer player, EnumHand hand) {
-        ItemStack itemstack = player.getHeldItem(hand);
-
-        if (!player.capabilities.isCreativeMode) itemstack.shrink(1);
-
-        world.playSound((EntityPlayer)null, player.posX, player.posY, player.posZ, SoundEvents.ENTITY_ENDERPEARL_THROW, SoundCategory.NEUTRAL, 0.5F, 0.4F / (itemRand.nextFloat() * 0.4F + 0.8F));
-        player.getCooldownTracker().setCooldown(this, 20);
-
-        if (!world.isRemote) {
-            EntityEMEnderPearl pearl = new EntityEMEnderPearl(world, player, itemstack);
-            pearl.shoot(player, player.rotationPitch, player.rotationYaw, 0.0F, 1.5F, 1.0F);
-            world.spawnEntity(pearl);
-        }
-
-        player.addStat(StatList.getObjectUseStats(this));
-        return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, itemstack);
-    }
 
 	@Override
 	public String texture() {
